Testing file and dir refactor in preparation for unit tests
- Created an e2e directory and placed in it all relevant files.
- Grouped e2e tests and named it Compiler.
- Minor refactor of test.js to satisfy linting
