+++ e2e/fi-samples/bool_test.fi
