+++ e2e/fi-samples/map_test.fi
