+++ e2e/fi-samples/string_test.fi
