+++ e2e/test.js
@@ -1,0 +1,43 @@
+'use strict'
+
+const fs = require('fs').promises
+const path = require('path')
+const ficompiler = require('../index')
+
+
+const compileFile = async file => {
+	const filePath = path.join(process.cwd(), file)
+	const data = await fs.readFile(filePath, {
+		encoding: 'utf-8'
+	})
+	return ficompiler.compile(data)
+}
+
+describe(`Compiler`, () => {
+	const contracts = [
+		'bigmap'
+		, 'blank'
+		, 'bool'
+		, 'bytes'
+		, 'input'
+		, 'key'
+		, 'key_hash'
+		, 'list'
+		, 'map'
+		, 'mutez'
+		, 'numbers'
+		, 'optional'
+		, 'set'
+		, 'signature'
+		, 'storage'
+		, 'string'
+		, 'timestamp'
+	]
+
+	contracts.forEach(contract => {
+		it(`${contract} sample compiles correctly`, async () => {
+			const compiled = await compileFile(`e2e/fi-samples/${contract}_test.fi`)
+			expect(compiled).toMatchSnapshot()
+		})
+	})
+})
