+++ jest.config.js
@@ -1,0 +1,10 @@
+'use strict'
+
+module.exports = {
+	coverageDirectory: 'e2e/coverage'
+	, collectCoverageFrom: [
+		'index.js'
+		, 'lib/**/*.js'
+	]
+	, verbose: true
+}
